Spec decode warmup support (#624)
GAUDISW-242931

Because currently spec decode flatten the spec decode tokens into
[batch_size * num_tokens, 1], we can warmup the decode shapes as it was.
The thing changed is the maximum batch_size we should warmup in the
configuration because the real batch size is batch_size * num_tokens
which is num_tokens (1 + num_speculative_tokens) times of original batch
size.

The thing to care in the warmup is the draft token (and block) space for
the proposing process in eagle. We need to leave out the
num_speculative_tokens space to use by propose for eagle.

Other care needs to be taken (already done in the PR of support
num_speculative_tokens > 1) is warmup will be run in compile only mode
without the real computation happening. So the operations for
prepare_attn_metadata in the drafter which depends on the real position
values must be done on CPU)

Another issue of handling no spec decode tokens for decode phase has
already been handled https://github.com/vllm-project/vllm-gaudi/pull/593

---------

Signed-off-by: Chen Haifeng <haifeng.chen@intel.com>